- Fix handling of HTTP Exceptions
- Refactor http reply handling
